Update tails.ipynb
attempting a fix for #437

i found a solution from https://github.com/jupyter-widgets/ipywidgets/issues/1267

there is not an easy way to test this. readthedocs does support multiple versions, but the repo isn't currently set up that way. 

it's also not possible to test this in the github editor, as the github editor already properly displays $.